completee v4l2 implementation start implelemt enumerate and UVCameraDevice

+++ src/util/camera/webcam.rs
@@ -3,9 +3,9 @@ use crate::util::camera::device::{DeviceHolder, Resolution};
 pub trait Webcam {
     fn name(&self) -> String;
     fn set_resolution(&mut self, res: Resolution);
-    fn set_framerate(&mut self, fps: i64);
+    fn set_framerate(&mut self, fps: u32);
     fn get_supported_resolutions(&self) -> Vec<Resolution>;
-    fn get_supported_framerate(&self, res: Resolution) -> Vec<i32>;
+    fn get_supported_framerate(&self, res: Resolution) -> Vec<u32>;
 }
 
 pub enum WebcamType {
